pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "libra"
version = "1.0.0"
authors = [
    { name = "Brady Spears" }
]
maintainers = [
    { name = 'Richard Alfaro-Diaz' },
    { name = 'Christine Gammans' },
    { name = 'Jonathan MacCarthy' }
]
description = "Your database manager's favority database manager"
readme = "README.md"
license = "MIT"
license-files = ["LICENSE.txt"]
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent"
]
dependencies = [
    "sqlalchemy>=2.0",
    "pyyaml"
]

[tool.setuptools.package-data]
"libra.schemas" = ["*.yaml"]

[project.optional-dependencies]
pandas = [
    "pandas"
]
rich = [
    "rich"
]
cx_oracle = [
    "cx_oracle"
]
all = [
    "pandas",
    "rich",
    "cx_oracle"
]

[project.urls]
Homepage = "https://github.com/LANL-Seismoacoustics/libra"